# api-collection/karnak-api.openapi.yaml
openapi: 3.0.3
info:
  title: Karnak API - ForwardNode, Profile, Project, Monitoring, Pseudonym, Echo
  description: >-
    OpenAPI 3.x specification covering the REST endpoints of the
    ForwardNodeController, ProfileController, ProjectController,
    MonitoringController, PseudonymController and EchoController controllers
    of Karnak.

    Authentication: Basic Auth (default username/password `admin`/`karnak`,
    see `application.yml` / `KARNAK_ADMIN` / `KARNAK_PASSWORD`).
  version: "1.0"
servers:
  - url: http://localhost:8081
    description: Default local instance (server.port in application.yml)
security:
  - basicAuth: []
...
paths:
...
  /api/echo/destinations:
    get:
      tags: [Echo]
      summary: Retrieve status and configured destinations
      description: Retrieve status and configured destinations from an AeTitle
      operationId: retrieveStatusConfiguredDestinations
      parameters:
        - name: srcAet
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status and configured destinations found from an AeTitle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationEchos"
            application/xml:
              schema:
                $ref: "#/components/schemas/DestinationEchos"
        "204":
          description: No Content
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    DestinationEcho:
      type: object
      description: Status and configuration of a single destination for an echo request.
      properties:
        aet:
          type: string
          description: AeTitle of the destination (dicom destinations only)
        url:
          type: string
          description: Url of the destination (stow destinations only)
        status:
          type: integer
          description: DICOM C-ECHO status (0 = success)
    DestinationEchos:
      type: object
      description: Wrapper list of destination echo results for a source AeTitle.
      properties:
        destinationEchos:
          type: array
          items:
            $ref: "#/components/schemas/DestinationEcho"